backend/requirements.txt
# --- Framework principal ---
fastapi==0.115.0
uvicorn==0.30.3
python-multipart==0.0.20

# --- ORM et Base de données ---
SQLAlchemy==2.0.32
pydantic==2.9.0
pydantic-settings==2.5.2

# --- Variables d'environnement ---
python-dotenv>=0.19.0

# --- Sécurité et Hashage ---
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# --- Gestion JWT ---
python-jose==3.3.0
cryptography==43.0.3

# --- Validation email ---
email-validator==2.1.1

# --- Tests (optionnel) ---
pytest==8.3.2
httpx==0.27.0

# --- Outils de développement (optionnel) ---
black==24.10.0
isort==5.13.2
